protecting some categories from being deleted

+++ app/Http/Controllers/CategoryController.php
@@ -94,8 +94,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $none = Category::where(['name'=>'None'])->first();
-        $category->dojos()->update(['category_id' => $none->id]);
-        $category->delete();
+        
+        if (auth()->user()->can('delete',$category)) {
+            $none = Category::where(['name'=>'None'])->first();
+            $category->dojos()->update(['category_id' => $none->id]);
+            $category->delete();
+        }
     }
 }
